implemented the app with react context api, seperated add todo to a separate component
