+++ src/store/TodoContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const TodoContext = createContext([]);
